kernel-module-mali: Switch to using community device-tree names
The community LIMA driver uses a series of names that are different then
the legacy Xilinx names.  Moves from the legacy names to the more standard
names.

This will allow us to have a single device tree that works with both the
mali driver, as well as the lima driver.

Interrupt-names:
  IRQGP     -> gp
  IRQGPMMU  -> gpmmu
  IRQPP0    -> pp0
  IRQPPMMU0 -> ppmmu0
  IRQPP1    -> pp1
  IRQPPMMU1 -> ppmmu1
  ...

Clock-names:
  gpu_pp0 -> core
  gpu_pp1 -> bus

Signed-off-by: Mark Hatle <mark.hatle@amd.com>